requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0

# Data Versioning
dvc>=3.0.0

# MLflow (tracking simple)
mlflow>=2.5.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Utils
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0


